backend/RAG/requirements.txt
flask
flask-cors
pymupdf
faiss-cpu
scikit-learn
google-generativeai
langchain
python-dotenv